refactor ipc files
