+++ package.json
@@ -19,6 +19,7 @@
     "main": "background.js",
     "dependencies": {
         "core-js": "^3.6.5",
+        "electron-store": "^8.0.0",
         "js-cookie": "^2.2.1",
         "mitt": "^3.0.0",
         "vue": "^3.0.0",
