+++ src/electron/pdf-events.js
